reorganize code for the typed choice sequence
