+++ hypothesis-python/src/hypothesis/internal/conjecture/engine.py
@@ -24,7 +24,7 @@ import attr
 
 from hypothesis import HealthCheck, Phase, Verbosity, settings as Settings
 from hypothesis._settings import local_settings
-from hypothesis.database import ExampleDatabase, ir_from_bytes, ir_to_bytes
+from hypothesis.database import ExampleDatabase, choices_from_bytes, choices_to_bytes
 from hypothesis.errors import (
     BackendCannotProceed,
     FlakyReplay,
@@ -38,18 +38,15 @@ from hypothesis.internal.conjecture.choice import (
     ChoiceKeyT,
     ChoiceKwargsT,
     ChoiceT,
+    ChoiceTemplate,
     choices_key,
 )
 from hypothesis.internal.conjecture.data import (
-    AVAILABLE_PROVIDERS,
     ConjectureData,
     ConjectureResult,
     DataObserver,
-    HypothesisProvider,
     IRNode,
-    NodeTemplate,
     Overrun,
-    PrimitiveProvider,
     Status,
     _Overrun,
 )
@@ -63,6 +60,11 @@ from hypothesis.internal.conjecture.junkdrawer import (
     startswith,
 )
 from hypothesis.internal.conjecture.pareto import NO_SCORE, ParetoFront, ParetoOptimiser
+from hypothesis.internal.conjecture.providers import (
+    AVAILABLE_PROVIDERS,
+    HypothesisProvider,
+    PrimitiveProvider,
+)
 from hypothesis.internal.conjecture.shrinker import Shrinker, ShrinkPredicateT, sort_key
 from hypothesis.internal.escalation import InterestingOrigin
 from hypothesis.internal.healthcheck import fail_health_check
@@ -193,10 +195,10 @@ StatisticsDict = TypedDict(
 )
 
 
-def choice_count(choices: Sequence[Union[ChoiceT, NodeTemplate]]) -> Optional[int]:
+def choice_count(choices: Sequence[Union[ChoiceT, ChoiceTemplate]]) -> Optional[int]:
     count = 0
     for choice in choices:
-        if isinstance(choice, NodeTemplate):
+        if isinstance(choice, ChoiceTemplate):
             if choice.count is None:
                 return None
             count += choice.count
@@ -361,7 +363,7 @@ class ConjectureRunner:
 
     def cached_test_function_ir(
         self,
-        choices: Sequence[Union[ChoiceT, NodeTemplate]],
+        choices: Sequence[Union[ChoiceT, ChoiceTemplate]],
         *,
         error_on_discard: bool = False,
         extend: Union[int, Literal["full"]] = 0,
@@ -376,9 +378,9 @@ class ConjectureRunner:
         """
         # node templates represent a not-yet-filled hole and therefore cannot
         # be cached or retrieved from the cache.
-        if not any(isinstance(choice, NodeTemplate) for choice in choices):
+        if not any(isinstance(choice, ChoiceTemplate) for choice in choices):
             # this type cast is validated by the isinstance check above (ie, there
-            # are no NodeTemplate elements).
+            # are no ChoiceTemplate elements).
             choices = cast(Sequence[ChoiceT], choices)
             key = self._cache_key(choices)
             try:
@@ -597,7 +599,7 @@ class ConjectureRunner:
             else:
                 if sort_key(data.nodes) < sort_key(existing.nodes):
                     self.shrinks += 1
-                    self.downgrade_buffer(ir_to_bytes(existing.choices))
+                    self.downgrade_choices(existing.choices)
                     self.__data_cache.unpin(self._cache_key(existing.choices))
                     changed = True
 
@@ -647,7 +649,7 @@ class ConjectureRunner:
         self.record_for_health_check(data)
 
     def on_pareto_evict(self, data: ConjectureData) -> None:
-        self.settings.database.delete(self.pareto_key, ir_to_bytes(data.choices))
+        self.settings.database.delete(self.pareto_key, choices_to_bytes(data.choices))
 
     def generate_novel_prefix(self) -> tuple[ChoiceT, ...]:
         """Uses the tree to proactively generate a starting sequence of bytes
@@ -738,9 +740,10 @@ class ConjectureRunner:
             key = self.sub_key(sub_key)
             if key is None:
                 return
-            self.settings.database.save(key, ir_to_bytes(choices))
+            self.settings.database.save(key, choices_to_bytes(choices))
 
-    def downgrade_buffer(self, buffer: Union[bytes, bytearray]) -> None:
+    def downgrade_choices(self, choices: Sequence[ChoiceT]) -> None:
+        buffer = choices_to_bytes(choices)
         if self.settings.database is not None and self.database_key is not None:
             self.settings.database.move(self.database_key, self.secondary_key, buffer)
 
@@ -854,7 +857,7 @@ class ConjectureRunner:
             for i, existing in enumerate(corpus):
                 if i >= primary_corpus_size and found_interesting_in_primary:
                     break
-                choices = ir_from_bytes(existing)
+                choices = choices_from_bytes(existing)
                 if choices is None:
                     # clear out any keys which fail deserialization
                     self.settings.database.delete(self.database_key, existing)
@@ -890,7 +893,7 @@ class ConjectureRunner:
                 pareto_corpus.sort(key=shortlex)
 
                 for existing in pareto_corpus:
-                    choices = ir_from_bytes(existing)
+                    choices = choices_from_bytes(existing)
                     if choices is None:
                         self.settings.database.delete(self.pareto_key, existing)
                         continue
@@ -957,7 +960,7 @@ class ConjectureRunner:
 
         assert self.should_generate_more()
         zero_data = self.cached_test_function_ir(
-            (NodeTemplate("simplest", count=None),)
+            (ChoiceTemplate("simplest", count=None),)
         )
         if zero_data.status > Status.OVERRUN:
             assert isinstance(zero_data, ConjectureResult)
@@ -1051,7 +1054,7 @@ class ConjectureRunner:
                 and consecutive_zero_extend_is_invalid < 5
             ):
                 minimal_example = self.cached_test_function_ir(
-                    prefix + (NodeTemplate("simplest", count=None),)
+                    prefix + (ChoiceTemplate("simplest", count=None),)
                 )
 
                 if minimal_example.status < Status.VALID:
@@ -1295,7 +1298,7 @@ class ConjectureRunner:
 
     def new_conjecture_data_ir(
         self,
-        prefix: Sequence[Union[ChoiceT, NodeTemplate]],
+        prefix: Sequence[Union[ChoiceT, ChoiceTemplate]],
         *,
         observer: Optional[DataObserver] = None,
         max_choices: Optional[int] = None,
@@ -1379,12 +1382,13 @@ class ConjectureRunner:
                 self.settings.database.fetch(self.secondary_key), key=shortlex
             )
             for c in corpus:
-                choices = ir_from_bytes(c)
+                choices = choices_from_bytes(c)
                 if choices is None:
                     self.settings.database.delete(self.secondary_key, c)
                     continue
                 primary = {
-                    ir_to_bytes(v.choices) for v in self.interesting_examples.values()
+                    choices_to_bytes(v.choices)
+                    for v in self.interesting_examples.values()
                 }
                 cap = max(map(shortlex, primary))
